backend/app/requirements.txt
fastapi
uvicorn[standard]
openai
langchain-community
langchain-openai
faiss-cpu
python-dotenv
langchain
pydantic
boto3
